+++ HISTORY.rst
@@ -5,6 +5,25 @@ History
 1.3dev
 ------
 
+- Add default css class for ``select`` blueprint.
+  [rnix, 2011-12-18]
+
+- Register ``number`` blueprint display renderer.
+  [rnix, 2011-12-18]
+
+- Consider ``expression`` in ``submit`` blueprint renderer.
+  [rnix, 2011-12-18]
+
+- ``checked`` attribute can be set explicitly in ``checkbox`` blueprint.
+  [rnix, 2011-11-21]
+
+- Fix Bug in ``yafowil.common.select_edit_renderer``. Crashed with empty
+  vocabularies.
+  [rnix, 2011-11-16]
+
+- Add ``lines`` blueprint. Renders a textarea and extracts lines as list.
+  [rnix, 2011-11-11]
+
 - Added concept of *plans* to the factory, which is a named set of blueprints.
   Plans are registered to the factory and can be addressed with the ``#`` sign.
   [jensens, 2011-09-29]